requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.11.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
newrelic==7.10.0.175
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
